backend/requirements.txt

# Azure OpenAI SDK
# Python 3.8以上が必要
# WebSocket通信
# Webフレームワーク
# 画像処理
# 設定管理
# データベース
# ファイル処理
# リクエスト処理
# バックエンド依存関係
Pillow==10.1.0            # 画像生成・処理用
aiofiles==23.2.1          # 非同期ファイル処理
aiohttp
aiohttp==3.12.13          # 非同期HTTPクライアント（画像生成用）pytz
beautifulsoup4
fastapi==0.104.1
httpx==0.25.2             # HTTPクライアント
openai>=1.99.0
pydantic-settings==2.1.0  # 設定スキーマ
pydantic==2.5.0           # データバリデーション
python-dotenv==1.0.0      # 環境変数管理
python-multipart==0.0.6    # ファイルアップロード用
python-socketio==5.10.0   # Socket.IO実装
pytz
tinydb==4.8.0             # 軽量NoSQLデータベース
uvicorn[standard]==0.24.0  # ASGIサーバー
websockets==12.0          # WebSocketクライアント/サーバー
